+++ rui-kotlin-plugin/src/main/kotlin/hu/simplexion/rui/kotlin/plugin/RuiConfigurationKeys.kt
@@ -9,9 +9,9 @@ import hu.simplexion.rui.runtime.Plugin.OPTION_NAME_DUMP_POINT
 import hu.simplexion.rui.runtime.Plugin.OPTION_NAME_EXPORT_STATE
 import hu.simplexion.rui.runtime.Plugin.OPTION_NAME_IMPORT_STATE
 import hu.simplexion.rui.runtime.Plugin.OPTION_NAME_PLUGIN_LOG_DIR
+import hu.simplexion.rui.runtime.Plugin.OPTION_NAME_PRINT_DUMPS
 import hu.simplexion.rui.runtime.Plugin.OPTION_NAME_ROOT_NAME_STRATEGY
 import hu.simplexion.rui.runtime.Plugin.OPTION_NAME_TRACE
-import hu.simplexion.rui.runtime.Plugin.OPTION_NAME_UNIT_TEST_MODE
 import org.jetbrains.kotlin.config.CompilerConfigurationKey
 
 object RuiConfigurationKeys {
@@ -21,7 +21,7 @@ object RuiConfigurationKeys {
     val TRACE: CompilerConfigurationKey<Boolean> = CompilerConfigurationKey.create(OPTION_NAME_TRACE)
     val EXPORT_STATE: CompilerConfigurationKey<Boolean> = CompilerConfigurationKey.create(OPTION_NAME_EXPORT_STATE)
     val IMPORT_STATE: CompilerConfigurationKey<Boolean> = CompilerConfigurationKey.create(OPTION_NAME_IMPORT_STATE)
-    val UNIT_TEST_MODE: CompilerConfigurationKey<Boolean> = CompilerConfigurationKey.create(OPTION_NAME_UNIT_TEST_MODE)
+    val PRINT_DUMPS: CompilerConfigurationKey<Boolean> = CompilerConfigurationKey.create(OPTION_NAME_PRINT_DUMPS)
     val PLUGIN_LOG_DIR: CompilerConfigurationKey<String> = CompilerConfigurationKey.create(OPTION_NAME_PLUGIN_LOG_DIR)
 }