+++ rui-kotlin-plugin/src/main/kotlin/hu/simplexion/rui/kotlin/plugin/globals.kt
@@ -74,6 +74,7 @@ const val RUI_SELECT = "ruiSelect"
 const val RUI_EXTERNAL_PATCH_OF_CHILD = "ruiEp"
 
 val RUI_FQN_FRAGMENT_CLASS = FqName.fromSegments(Plugin.RUI_FRAGMENT_CLASS)
+val RUI_FQN_GENERATED_FRAGMENT_CLASS = FqName.fromSegments(Plugin.RUI_GENERATED_FRAGMENT_CLASS)
 val RUI_FQN_ADAPTER_CLASS = FqName.fromSegments(Plugin.RUI_ADAPTER_CLASS)
 val RUI_FQN_BRIDGE_CLASS = FqName.fromSegments(Plugin.RUI_BRIDGE_CLASS)
 val RUI_FQN_BLOCK_CLASS = FqName.fromSegments(Plugin.RUI_BLOCK_CLASS)