+++ doc/Changelog.md
@@ -1,3 +1,12 @@
+# 0.2.0-SNAPSHOT
+
+- Issue #4 - Use RuiGeneratedFragment instead of RuiFragment in generated classes
+- run and success unit tests now generate code with and without trace
+- change `unitTestMode` plugin option to `printDumps`
+- add `jvmToolchain(11)` to projects
+- make class `RuiDOMAdapter` open
+- convert runtime manual block and branch test into a proper unit test
+
 # 0.1.0
 
 Works: